fix: 模型接入、改为右手系

+++ KairosEngine/src/graphics/mesh.rs
@@ -6,11 +6,11 @@ use crate::graphics::vertex::Vertex;
 pub struct Mesh {
     pub id: usize,
     pub vertices: Vec<Vertex>,
-    pub indices: Vec<u32>,
+    pub indices: Vec<u16>,
 }
 
 impl Mesh {
-    pub fn new(id: usize, vertices: Vec<Vertex>, indices: Vec<u32>) -> Self {
+    pub fn new(id: usize, vertices: Vec<Vertex>, indices: Vec<u16>) -> Self {
         Self {
             id,
             vertices,
